Add prop-types as peer-dependency
This library currently depend implicitly on prop-types being installed and if it's not implicitly installed, then it fails. That's bad. We *need* to either add it as a dependency or a peerDependency. Please don't get this PR caught up in a bureaucratic clusterf#@%. I have no opinion on how this should be done. But it *needs* to be done.