Correct sharing a video with preview photo (#555)
Previously this feature was removed from iOS and JS side of the library. It turned out that it is supported by the native FBSDK, so this PR adds it back in, the correct way.

As an extra I've fixed the iOS build issue of the example app.

The example app builds and runs.