Update README.md (#731)
https://developers.facebook.com/docs/ios/getting-started is confusing for us because:

'In the following list, select the appropriate SDK for the language you will use in this app. If your app uses Swift, select SDK: Cocoapods. If your app uses Objective-C, select SDK: FB SDK'

None of these are true in our case, but we do want to have 'SDK: Cocoapods' pre-selected.

Further, 1.1 clarifies to only go as far as '1. Install Cocoapods.' and not to install pod 'FBSDKCoreKit/Swift'.